requirements.txt
Django==2.0.2
psycopg2-binary==2.7.4
django-debug-toolbar==1.9.1
django-autofixture==0.12.1
social-auth-app-django==2.1.0
social-auth-core==1.7.0
djangorestframework==3.7.7

